Implemented PlantSystem.
Plant System now handles properties on a level higher than the individual, like positioning the plants.
More importantly, the vertex data of all plants is now contained in big, shared buffers.
The update on the computer shader is now done in a single call.
The extra buffer for vertex drawing was discarded. Instead, the dynamic buffer is reused.
Preparing the dynamic buffer for drawing is still done element by element due to parent dependency.

Small change in text drawing to account for screen size.